third_party/WebKit/LayoutTests/fast/forms/color/display-none-input-color-chooser-shown.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/color_chooser/color_chooser.mojom.js"></script>
<script src="../../forms/color/mock-colorchooser.js"></script>
<input id="colorPick" type="color" />
<label for="colorPick" id="labelPick">Pick a color</label>
<script>
test (function() {
    assert_true(window.eventSender !== null);
}, "window.eventSender is required for the test to run");

async_test (t => {
    var colorPicker = document.getElementById("colorPick");
    colorPicker.style.display = "none";
    var labelPick = document.getElementById("labelPick");
    var x = labelPick.offsetLeft + labelPick.offsetWidth/2;
    var y = labelPick.offsetTop + labelPick.offsetHeight/2;
    eventSender.mouseMoveTo(x, y);
    eventSender.mouseDown();
    eventSender.mouseUp();
    waitUntilChooserShown(() => { t.done(); });
}, "Tests click on label for color picker with display none should show chooser");
</script>
